Refacto tile coordinate system and add isometric coordinates

+++ src/Assets/Definitions.cpp
@@ -1,8 +1,8 @@
 #include "Definitions.h"
 
-std::pair<int, int> Definitions::orthoToIso(std::pair<int, int> orthoCoords)
+std::pair<double, double> Definitions::orthoToIso(std::pair<int, int> orthoCoords)
 {
 	std::pair<int, int> coordinates = { spriteSize * orthoCoords.first, spriteSize * orthoCoords.second };
-	std::pair<int, int> isoCoords = { (coordinates.first - coordinates.second) * 0.5, (coordinates.first + coordinates.second) * 0.25 };
+	std::pair<double, double> isoCoords = { (coordinates.first - coordinates.second) * 0.5, (coordinates.first + coordinates.second) * 0.25 };
 	return isoCoords;
 }
